move relevant methods to _integrations namespace
